# libc: arm64: add optimized strcpy implementation