This optimization is extracted from cortex-strings and bionic-ized.
stringbench result:
|---------------+--------+--------+-------------+
|   item        |  aosp  | cortex | cortex gain |
|---------------+--------+--------+-------------+
| strcpy 15000b | 5.23 s |  5.20s |       0.46% |
|---------------+--------+--------+-------------+
| strcpy 15b    | 3.64 s |  3.28s |       9.91% |
|---------------+--------+--------+-------------+

Change-Id: If04ebf677f8f41b4555b00709d41cc09b11cc927
Signed-off-by: Yingshiuan Pan <yingshiuan.pan@linaro.org>